Changed the pattern/code setup
All graphs and discussion about data quality is now hidden away until
the click on a button. The underlying code is not dramatically changed,
so now survey data is loaded until it is actually needed. Good for
performance and also nice visually, as the user now gets the chance to
see each graph’s loading animation.